dom/tests/mochitest/general/test_bug631440.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=631440
-->
<head>
  <title>Test for Bug 631440</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=631440">Mozilla Bug 631440</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

SimpleTest.waitForExplicitFinish();

var w = window.open("about:blank");
w.addEventListener("load", function() {
  w.close();
  SimpleTest.finish();
}, false);

try {
  w.history.pushState(null, "title", "pushState.html");
  ok(false, "Should have thrown a security exception");
}
catch (e) {
  ok(true, "Should have thrown a security exception");
}

</script>
</pre>
</body>
</html>
